refactor: move pool/hooks.ts
