+++ src/app/components/trade/LiquidityDetails.tsx
@@ -22,9 +22,9 @@ import { Typography } from 'app/theme';
 import bnJs from 'bnJs';
 import { NETWORK_ID } from 'constants/config';
 import { BIGINT_ZERO, FRACTION_ONE } from 'constants/misc';
+import { useBalance, usePool, usePoolData, useAvailableBalances, pairToken } from 'hooks/usePools';
 import { useChangeShouldLedgerSign, useShouldLedgerSign } from 'store/application/hooks';
 import { Field } from 'store/mint/actions';
-import { useBalance, usePool, usePoolData, useAvailableBalances, pairToken } from 'store/pool/hooks';
 import { tryParseAmount } from 'store/swap/hooks';
 import { useTransactionAdder } from 'store/transactions/hooks';
 import { useCurrencyBalances, useHasEnoughICX } from 'store/wallet/hooks';
